improvement(3.6): auto resize text area
I first tried following the guide on Geeks for Geeks website, but it did not produce desired results. Then I discovered the guide on Angular Material website via answer on Stack Overflow website.
Additional improvement was to only set the values of form controls when they are updated with fetched data from server. There is no need to override validators, so I am calling the method "setValue" on form.

References:
https://www.geeksforgeeks.org/how-to-create-auto-resize-textarea-using-javascript-jquery/

https://v12.material.angular.io/cdk/text-field/overview (via https://stackoverflow.com/a/65306162)

https://stackoverflow.com/a/55275042